validacoes adicionadas

+++ BibliotecaController/src/br/com/anhanguera/pos/biblioteca/controller/ExemplarController.java
@@ -10,16 +10,27 @@ import java.util.List;
  */
 public class ExemplarController {
 
-    private Exemplar exemplar;
+    public static String msg = "O campo {CAMPO} é obrigatório.";
     
-    public ExemplarController(Exemplar exemplar){
-        this.exemplar = exemplar;
+    private static boolean isValid(Exemplar e, boolean bValidaCodigo){
+        if(bValidaCodigo && e.getCodigoExemplar() == 0)
+        { msg.replace("{CAMPO}", "Código"); return false; }
+        else if (e.getObra().getCodigoObra() == 0)
+        { msg.replace("{CAMPO}", "Obra"); return false; }
+        else if (e.getSituacaoExemplar().equals(null))
+        { msg.replace("{CAMPO}", "Situação"); return false; }
+        else
+            return false;
     }
-    
+
     public ExemplarController(){ }
     
-    public boolean insert(){
+    public boolean insert(Exemplar exemplar){
         try{
+            
+            if(!isValid(exemplar, false))
+                return false;
+            
             return new ExemplarDAO().insert(exemplar);
         }catch(Exception e){
             System.out.println(e.getMessage());
@@ -27,8 +38,12 @@ public class ExemplarController {
         }
     }
     
-    public boolean alter(){
+    public boolean alter(Exemplar exemplar){
         try{
+            
+            if(!isValid(exemplar, true))
+                return false;
+            
             return new ExemplarDAO().alter(exemplar);
         }catch(Exception e){
             return false;
@@ -37,6 +52,10 @@ public class ExemplarController {
     
     public boolean delete(int intCodigoExemplar, int intCodigoObra){
         try{
+            
+            if(intCodigoExemplar == 0 || intCodigoObra == 0)
+                return false;
+            
             return new ExemplarDAO().delete(intCodigoExemplar, intCodigoObra);
         }catch(Exception e){
             return false;
@@ -59,7 +78,7 @@ public class ExemplarController {
         }
     }
     
-    public List<Exemplar> select(){
+    public List<Exemplar> select(Exemplar exemplar){
         try{
             return new ExemplarDAO().select(exemplar);
         }catch(Exception e){
